pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "publiminer"
version = "0.1.0"
description = "Publication Literature Miner — PubMed API, embedding-based clustering, and LLM structured-output extraction"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Seyed Amir Ahmad Safavi-Naini" },
]
keywords = ["pubmed", "literature-mining", "nlp", "clustering", "llm", "bioinformatics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "polars>=1.0",
    "pyarrow>=14.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "lxml>=5.0",
    "typer>=0.12",
    "rich>=13.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "hdbscan>=0.8",
    "thefuzz[speedup]>=0.22",
    "jinja2>=3.1",
]

[project.optional-dependencies]
viz = [
    "umap-learn>=0.5",
    "matplotlib>=3.8",
    "plotly>=5.18",
]
rag = [
    "chromadb>=0.4",
]
ui = [
    "streamlit>=1.30",
    "openpyxl>=3.1",
    "xlsxwriter>=3.2",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
all = ["publiminer[viz,rag,ui,dev]"]

[project.urls]
Repository = "https://github.com/sdamirsa/PubLiMiner"
Issues = "https://github.com/sdamirsa/PubLiMiner/issues"

[project.scripts]
publiminer = "publiminer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/publiminer"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
